fix(lua): fix before merging master

+++ src/lua/bindings/native.hpp
@@ -14,7 +14,7 @@ namespace lua::native
 		for (int i = 1; i < args.size(); i++)
 		{
 			if (args[i].is<int>())
-				big::g_native_invoker.push_arg(args[i].get<int>());
+				big::g_native_invoker.push_arg(args[i].get<uint64_t>());
 			if (args[i].is<float>())
 				big::g_native_invoker.push_arg(args[i].get<float>());
 			else if (args[i].is<bool>())
